tGreenplumGPLoad/tGreenplumGPLoad_begin.javajet / tGreenplumGPLoad/tGreenplumGPLoad_java.xml Added guessColumnList so the yaml will contain the columns of the talend schema, this is extremely useful if the target table has extra columns (for example to load only the primary keys to a staging table) This will generate the "  COLUMNS:" section of the YAML file.
Added parameters to the advanced settings, more information for these can be found at: https://gpdb.docs.pivotal.io/6-12/utility_guide/ref/gpload.html#topic1__section7
Parameters include:
  MAX_LINE_LENGTH:
  EXTERNAL:
    -SCHEMA:
  PRELOAD:
    -TRUNCATE:
    -REUSE_TABLES:
    -STAGING_TABLE:
    -FAST_MATCH:
  SQL:
    -BEFORE:
    -AFTER:

Changed the way command is built, so the Log4j now displays a command that can be copy-pasted to the console for testing.
The yaml file was missing from it.
The PythonPath was also not printed to the log4j logs.

tGreenplumGPLoad/tGreenplumGPLoad_main.javajet
Added byte[] type support. Byte[] is converted to escaped OCTET characters, which is compatible with all PostgreSQL versions from 8 to 13.

Fixed when newline / escape characters appeared in the data.

Postgresql uses null terminated bytes so those have to be removed from the data to avoid errors.